okay as far as i can tell I have done everything I set out to do today
leaderbaord is fully working

the logout button does google signout signs out, i think, its hard to test this one, but it should be.

when you press the exercise button it uploads it to the database which means in real time next time you go to leaderboard it will reupdate and your progress will be different

It handles mulitple of the same value scores in the leaderboard successfully (i.e if there are 4 4's and 1 it will show the 4's and then the 1.

New users are created when submitting the first exercise

if a user matching your userid stored in shared preference already exists, then it will simply update yours

(if you wiped your app, we currently have no way to re-establish a link but I dont care, major game apps have had that issue and worse so fuck it)

i MIGHT have fixed google SSO for you, but again hard to tell since mine has worked in all cases 

you might need to confirm its that, it imght auto update, im not sure.  The reason it might update is because i created a debug keystore now that is suppose to store it, but fucked if i know if it will.

there is currently no way to reset your exercise score except manually through the database, which is here

https://console.firebase.google.com/u/0/project/fitness-app-7d01d/database/fitness-app-7d01d-default-rtdb/data

but thats because the code was commented out

i added to the shared prefernces a first and lastname for Dev Account if you press the not google sso button in login